Convert weight and related fields to BigDecimal.
Refactored weight, gravity, and cargo capacity fields in multiple models to use BigDecimal for increased precision. Added backward-compatible setters and constructors to support legacy double-based code. Updated comparison logic to accommodate BigDecimal operations where applicable.

+++ api/src/main/java/com/barrows/travller/api/model/Weapon.java
@@ -2,6 +2,7 @@ package com.barrows.travller.api.model;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 import jakarta.persistence.*;
 
 /**
@@ -69,7 +70,7 @@ public class Weapon {
     /**
      * The weight of the weapon in kg.
      */
-    private double weight;
+    private BigDecimal weight;
 
     /**
      * The skill used with this weapon.
